MySQL fix + new feature
Added the feature to delete all notifications in the database when a player leave or keep them, also fixed a problem in MySQL where the created table wouldn't set the id default value as AUTO_INCREMENT.

+++ src/CupidonSauce173/PigNotify/NotifLoader.php
@@ -177,6 +177,10 @@ class NotifLoader extends PluginBase implements Listener
     {
         $player = $event->getPlayer();
         if (!isset($this->notificationList[$player->getName()])) return;
-        $this->deleteNotifications($this->getPlayerNotifications($player->getName()));
+        if($this->config['delete-on-disconnect']){
+            $this->deleteNotifications($this->getPlayerNotifications($player->getName()));
+            return;
+        }
+        unset($this->notificationList[$player->getName()]);
     }
 }
